+++ src/NewtonSolver.cpp
@@ -4,7 +4,7 @@ NewtonSolver::NewtonSolver(const arma::mat &X,
                            const arma::vec &y,
                            const arma::vec &o,
                            double epsilon,
-                           uint32_t maxIter,
+                           uint64_t maxIter,
                            double sigma) {
     setSolver(X, y, o, epsilon, maxIter, sigma);
 }
@@ -13,7 +13,7 @@ void NewtonSolver::setSolver(const arma::mat &X,
                            const arma::vec &y,
                            const arma::vec &o,
                            double epsilon,
-                           uint32_t maxIter,
+                           uint64_t maxIter,
                            double sigma) {
     this->X = X;
     this->y = y;
@@ -48,7 +48,7 @@ arma::vec NewtonSolver::solve(const arma::mat &X,
                     const arma::vec &y,
                     const arma::vec &o,
                     double epsilon,
-                    uint32_t maxIter,
+                    uint64_t maxIter,
                     double sigma){
     setSolver(X, y, o, epsilon, maxIter, sigma);
     return solve();
@@ -59,20 +59,20 @@ void NewtonSolver::setEpsilon(const double epsilon) {
 }
 
 arma::vec NewtonLogistic::getGradient(const arma::vec &beta) {
-    uint32_t n = X.n_rows, p = X.n_cols;
+    uint64_t n = X.n_rows, p = X.n_cols;
     arma::vec vecP(n);
-    for (uint32_t i = 0; i < n; i++) {
+    for (uint64_t i = 0; i < n; i++) {
         vecP[i] = 1 / (1 + exp(-(o[i] + arma::as_scalar(X.row(i) * beta))));
     }
     return X.t() * (vecP - y) / n + epsilon * beta;
 }
 
 arma::mat NewtonLogistic::getHessian(const arma::vec &beta) {
-    uint32_t n = X.n_rows, p = X.n_cols;
+    uint64_t n = X.n_rows, p = X.n_cols;
     double doubleP;
     arma::mat W(n, n, arma::fill::zeros);
     
-    for (uint32_t i = 0; i < n; i++) {
+    for (uint64_t i = 0; i < n; i++) {
         doubleP = 1 / (1 + exp(-(o[i] + arma::as_scalar(X.row(i) * beta))));
         W(i, i) = doubleP * (1 - doubleP);
     }
@@ -80,7 +80,7 @@ arma::mat NewtonLogistic::getHessian(const arma::vec &beta) {
 }
 
 arma::vec NewtonLogistic::solve() {
-    uint32_t k = 0;
+    uint64_t k = 0;
     if (X.empty() || y.empty()) {
         Rcpp::stop("The input matrix and response vector shouldn't be empty!");
     }
@@ -97,27 +97,27 @@ arma::vec NewtonLogistic::solve() {
 }
 
 arma::vec NewtonPoisson::getGradient(const arma::vec &beta) {
-    uint32_t n = X.n_rows, p = X.n_cols;
+    uint64_t n = X.n_rows, p = X.n_cols;
     arma::vec vecV(n);
 
-    for (uint32_t i = 0; i < n; i++) {
+    for (uint64_t i = 0; i < n; i++) {
         vecV[i] = exp(o[i] + arma::as_scalar(X.row(i) * beta));
     }
     return X.t() * (vecV - y) / n + epsilon * beta;
 }
 
 arma::mat NewtonPoisson::getHessian(const arma::vec &beta) {
-    uint32_t n = X.n_rows, p = X.n_cols;
+    uint64_t n = X.n_rows, p = X.n_cols;
     arma::mat W(n, n, arma::fill::zeros);
 
-    for (uint32_t i = 0; i < n; i++) {
+    for (uint64_t i = 0; i < n; i++) {
         W(i, i) = exp(o[i] + arma::as_scalar(X.row(i) * beta));
     }
     return X.t() * W * X / n + epsilon * arma::eye<arma::mat>(p, p);
 }
 
 arma::vec NewtonPoisson::solve() {
-    uint32_t k = 0;
+    uint64_t k = 0;
     if (X.empty() || y.empty()) {
         Rcpp::stop("The input matrix and response vector shouldn't be empty!");
     }
@@ -137,6 +137,6 @@ arma::vec NewtonGaussian::solve() {
     if (X.empty() || y.empty()) {
         Rcpp::stop("The input matrix and response vector shouldn't be empty!");
     }
-    uint32_t n = X.n_rows, p = X.n_cols;
+    uint64_t n = X.n_rows, p = X.n_cols;
     return (X.t() * X / n + epsilon * arma::eye<arma::mat>(p, p)).i() * X.t() * (y - o) / n;
 }
