+++ src/NewtonSolver.h
@@ -9,7 +9,7 @@ protected:
     arma::vec y;
     arma::vec o;
     double epsilon;
-    uint32_t maxIter;
+    uint64_t maxIter;
     double sigma;  // The convergence threshhold
 public:
     NewtonSolver(){};
@@ -17,13 +17,13 @@ public:
                  const arma::vec &y,
                  const arma::vec &o = arma::vec(),
                  double epsilon = 0.25,
-                 uint32_t maxIter = 1e5,
+                 uint64_t maxIter = 1e5,
                  double sigma = 1e-5);
     void setSolver(const arma::mat &X,
                    const arma::vec &y,
                    const arma::vec &o = arma::vec(),
                    double epsilon = 0.25,
-                   uint32_t maxIter = 1e5,
+                   uint64_t maxIter = 1e5,
                    double sigma = 1e-5);
     arma::vec fit(std::string method);
     virtual arma::vec solve(){};
@@ -31,7 +31,7 @@ public:
                     const arma::vec &y,
                     const arma::vec &o = arma::vec(),
                     double epsilon = 0.25,
-                    uint32_t maxIter = 1e5,
+                    uint64_t maxIter = 1e5,
                     double sigma = 1e-5);
     void setEpsilon(const double epsilon = 0.25);
     virtual ~NewtonSolver(){};
